check repos

+++ repos/allocation_repo.py
@@ -50,6 +50,11 @@ async def get_allocations(app: FastAPI):
 
 async def update_allocation(app: FastAPI, allocation_id: str, allocation: AllocationSchema):
     try:
+        if not (await check_existance_employee(app, allocation.employee_id)):
+            logger.error(f"Employee does not exist")
+            raise HTTPException(status_code=400, detail="Employee does not exist")
+        if allocation.allocated_date <= date.today():
+            raise HTTPException(status_code=400, detail="Allocated date cannot be in the past")
         allocation_dict = allocation.model_dump()
         result = await app.mongodb[ALLOCATION_COLLECTION].update_one({"_id": ObjectId(allocation_id)}, {"$set": allocation_dict})
     except Exception as e:
@@ -61,6 +66,9 @@ async def update_allocation(app: FastAPI, allocation_id: str, allocation: Alloca
 
 async def delete_allocation(app: FastAPI, allocation_id: str):
     try:
+        allocation_object = await app.mongodb[ALLOCATION_COLLECTION].find_one({"_id": ObjectId(allocation_id)})
+        if date.today() >= allocation_object.allocated_date:
+            return False
         result = await app.mongodb[ALLOCATION_COLLECTION].delete_one({"_id": ObjectId(allocation_id)})
     except Exception as e:
         logger.error(f"Error deleting allocation: {e}")
